backend/requirements.txt
django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.9
python-jose==3.3.0
boto3==1.38.21
pytest==8.3.5
pytest-django==4.11.1
gunicorn==21.2.0 